add auto generate id

+++ backend/src/main/java/com/duanWebChat/WebChatApplication/repository/MessageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.duanWebChat.WebChatApplication.entity.message.Message;
 
-public interface MessageRepository extends MongoRepository<Message, Integer>{
+public interface MessageRepository extends MongoRepository<Message, Long>{
 
 }
