+++ backend/src/main/java/com/duanWebChat/WebChatApplication/repository/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import com.duanWebChat.WebChatApplication.entity.user.User;
 
 
-public interface UserRepository extends MongoRepository<User, Integer>{
+public interface UserRepository extends MongoRepository<User, Long>{
 
 }
